feat!: merge discover_trending and recommend_content into discover
