+++ src/tools/implementations/content-discovery-tools.ts
@@ -17,17 +17,6 @@ const FindSimilarUsersSchema = z.object({
 });
 
 /**
- * Zod schema for recommend content parameters
- */
-const RecommendContentSchema = z.object({
-  maxResults: z.number().int().min(1).max(100).optional().default(20),
-  minLikes: z.number().int().min(0).optional().default(5),
-  maxAge: z.number().int().min(1).max(168).optional().default(24), // hours
-  topics: z.array(z.string()).optional(),
-  excludeReposts: z.boolean().optional().default(false),
-});
-
-/**
  * Zod schema for discover communities parameters
  */
 const DiscoverCommunitiesSchema = z.object({
@@ -359,246 +348,6 @@ export class FindSimilarUsersTool extends BaseTool {
         this.logger.warn('Candidate profile hydration failed for a chunk', error as Error);
       }
     }
-  }
-}
-
-/**
- * Tool for recommending content based on user interests
- */
-export class RecommendContentTool extends BaseTool {
-  public readonly schema = {
-    method: 'recommend_content',
-    description:
-      'Recommend posts based on user interests and engagement history. ' +
-      'Analyzes your timeline, engagement patterns, and network to suggest relevant content.',
-    params: RecommendContentSchema,
-  };
-
-  constructor(atpClient: AtpClient) {
-    super(atpClient, 'RecommendContent', ToolAuthMode.PRIVATE);
-  }
-
-  protected async execute(params: z.infer<typeof RecommendContentSchema>): Promise<{
-    success: boolean;
-    recommendations: Array<{
-      uri: string;
-      cid: string;
-      author: {
-        did: string;
-        handle: string;
-        displayName?: string;
-        avatar?: string;
-      };
-      text: string;
-      likeCount: number;
-      replyCount: number;
-      repostCount: number;
-      indexedAt: string;
-      recommendationScore: number;
-      recommendationReasons: string[];
-      topics?: string[];
-    }>;
-    insights: string[];
-  }> {
-    try {
-      this.logger.info('Generating content recommendations', {
-        maxResults: params.maxResults,
-        topics: params.topics,
-      });
-
-      const agent = this.atpClient.getAgent();
-
-      // Get user's timeline to understand their network
-      const timelineResponse = await this.executeAtpOperation(
-        async () => agent.getTimeline({ limit: 100 }),
-        'getTimeline',
-        { limit: 100 }
-      );
-
-      // The repost indicator lives on the feed item's `reason`
-      // (app.bsky.feed.defs#reasonRepost), not on the post record — preserve it so
-      // excludeReposts can actually filter reposts.
-      const timelinePosts = timelineResponse.data.feed.map((item: any) => ({
-        ...item.post,
-        __isRepost: item.reason?.$type === 'app.bsky.feed.defs#reasonRepost',
-      }));
-
-      // Get user's recent likes to understand preferences
-      const likedTopics = new Set<string>();
-      const likedAuthors = new Set<string>();
-
-      try {
-        await this.executeAtpOperation(
-          async () => agent.getProfile({ actor: agent.session?.did || '' }),
-          'getProfile',
-          {}
-        );
-
-        // Note: AT Protocol doesn't have a direct "get my likes" endpoint
-        // We'll infer from timeline engagement instead
-        for (const post of timelinePosts) {
-          if (post.viewer?.like) {
-            likedAuthors.add(post.author.did);
-            const topics = this.extractTopics([post]);
-            topics.forEach(t => likedTopics.add(t));
-          }
-        }
-      } catch (error) {
-        this.logger.warn('Could not analyze user preferences', error);
-      }
-
-      // Filter and score posts
-      const now = new Date();
-      const maxAgeMs = params.maxAge * 60 * 60 * 1000;
-      const recommendations = [];
-
-      for (const post of timelinePosts) {
-        const postData = post;
-
-        // Skip if already liked
-        if (postData.viewer?.like) continue;
-
-        // Skip reposts if requested (flag derived from the feed item's reason).
-        if (params.excludeReposts && postData.__isRepost) continue;
-
-        // Check age
-        const postAge = now.getTime() - new Date(postData.indexedAt).getTime();
-        if (postAge > maxAgeMs) continue;
-
-        // Check minimum likes
-        const likeCount = postData.likeCount || 0;
-        if (likeCount < params.minLikes) continue;
-
-        // Extract topics from post
-        const postTopics = Array.from(this.extractTopics([post]));
-
-        // Check topic filter. Match each requested topic against BOTH the post's
-        // hashtags and its text body — most Bluesky posts have no hashtags, so a
-        // hashtag-only filter dropped the majority of genuinely on-topic posts.
-        if (params.topics && params.topics.length > 0) {
-          const postText = (postData.record?.text || '').toLowerCase();
-          const hasMatchingTopic = params.topics.some(filter => {
-            const f = filter.toLowerCase();
-            return postTopics.some(topic => topic.includes(f)) || postText.includes(f);
-          });
-          if (!hasMatchingTopic) continue;
-        }
-
-        // Calculate recommendation score
-        let score = 0;
-        const reasons: string[] = [];
-
-        // Engagement score
-        const engagementScore =
-          likeCount * 1 + (postData.replyCount || 0) * 2 + (postData.repostCount || 0) * 1.5;
-        score += Math.min(engagementScore, 100);
-        if (likeCount >= params.minLikes * 2) {
-          reasons.push(`High engagement (${likeCount} likes)`);
-        }
-
-        // Author preference
-        if (likedAuthors.has(postData.author.did)) {
-          score += 30;
-          reasons.push('From an author you frequently engage with');
-        }
-
-        // Topic relevance
-        const topicMatches = postTopics.filter(t => likedTopics.has(t)).length;
-        if (topicMatches > 0) {
-          score += topicMatches * 20;
-          reasons.push(`Matches ${topicMatches} of your interests`);
-        }
-
-        // Recency bonus
-        const ageHours = postAge / (60 * 60 * 1000);
-        if (ageHours < 6) {
-          score += 10;
-          reasons.push('Recent post');
-        }
-
-        // Thread bonus (replies often have good discussions)
-        if (postData.replyCount && postData.replyCount > 3) {
-          score += 15;
-          reasons.push('Active discussion');
-        }
-
-        if (reasons.length === 0) {
-          reasons.push('Popular in your network');
-        }
-
-        recommendations.push({
-          uri: postData.uri,
-          cid: postData.cid,
-          author: {
-            did: postData.author.did,
-            handle: postData.author.handle,
-            displayName: postData.author.displayName,
-            avatar: postData.author.avatar,
-          },
-          text: postData.record?.text || '',
-          likeCount: postData.likeCount || 0,
-          replyCount: postData.replyCount || 0,
-          repostCount: postData.repostCount || 0,
-          indexedAt: postData.indexedAt,
-          recommendationScore: Math.round(score),
-          recommendationReasons: reasons,
-          topics: postTopics.length > 0 ? postTopics : undefined,
-        });
-      }
-
-      // Sort by recommendation score
-      recommendations.sort((a, b) => b.recommendationScore - a.recommendationScore);
-      const topRecommendations = recommendations.slice(0, params.maxResults);
-
-      // Generate insights
-      const insights: string[] = [];
-      if (topRecommendations.length > 0) {
-        insights.push(`Found ${topRecommendations.length} recommended posts from your network`);
-
-        const avgScore =
-          topRecommendations.reduce((sum, r) => sum + r.recommendationScore, 0) /
-          topRecommendations.length;
-        insights.push(`Average recommendation score: ${avgScore.toFixed(1)}`);
-
-        const topAuthors = new Set(topRecommendations.slice(0, 5).map(r => r.author.handle));
-        insights.push(`Top authors: ${Array.from(topAuthors).join(', ')}`);
-
-        const allTopics = new Set<string>();
-        topRecommendations.forEach(r => r.topics?.forEach(t => allTopics.add(t)));
-        if (allTopics.size > 0) {
-          insights.push(`Common topics: ${Array.from(allTopics).slice(0, 5).join(', ')}`);
-        }
-      } else {
-        insights.push('No recommendations found matching your criteria');
-        insights.push(
-          'Try adjusting filters (lower minLikes, increase maxAge, or remove topic filters)'
-        );
-      }
-
-      this.logger.info('Content recommendations generated', {
-        count: topRecommendations.length,
-      });
-
-      return {
-        success: true,
-        recommendations: topRecommendations,
-        insights,
-      };
-    } catch (error) {
-      this.logger.error('Failed to generate recommendations', error);
-      this.formatError(error);
-    }
-  }
-
-  private extractTopics(posts: any[]): Set<string> {
-    const topics = new Set<string>();
-    for (const post of posts) {
-      const text = post.record?.text || '';
-      // Extract hashtags
-      const hashtags = text.match(/#\w+/g) || [];
-      hashtags.forEach((tag: string) => topics.add(tag.toLowerCase()));
-    }
-    return topics;
   }
 }
